+++ test/runtests.jl
@@ -57,6 +57,7 @@ function include_fast_tests()
   include("test_island_diagnostics.jl")
   include("test_short_circuit.jl")
   include("test_external_grid.jl")
+  include("test_parallel_foundation.jl")
 end
 
 function include_extended_tests()
@@ -108,6 +109,7 @@ function run_fast_profile_tests()
     ("island_diagnostics", () -> run_entry(:run_island_diagnostics_tests)),
     ("short_circuit", () -> run_entry(:run_short_circuit_tests)),
     ("external_grid", () -> run_entry(:run_external_grid_tests)),
+    ("parallel_foundation", () -> run_entry(:run_parallel_foundation_tests)),
     ("controls", () -> begin
       run_entry(:run_voltage_dependent_control_tests)
       run_entry(:run_transformer_phase_shift_tests)